Smooth sidebar sticky header swaps
